Generate Command Text to Start the TiddlyWiki Server

This internal function constructs the command text needed to start the TiddlyWiki server based on the provided parameters. The generated command text can be used to start the server with specific configurations.

Usage

make_run_txt(tidepath = "myTide", mws = FALSE, tiddlywiki = NA, wikipath = NA)

Arguments

tidepath

Character. The path where the tide folder is located. Default is "myTide".

mws

Logical. Indicates whether to set up the TiddlyWiki server in Multi-User Server mode. Default is FALSE.

tiddlywiki

Character. The path to the TiddlyWiki installation. Default is NA.

wikipath

Character. The path where the wiki data will be stored. Default is NA.

Value

A character string that contains the command to start the TiddlyWiki server with the specified settings.

Examples

# Example usage:
make_run_txt()
#> Error in make_run_txt(): could not find function "make_run_txt"
make_run_txt(tidepath = "customPath", mws = TRUE, tiddlywiki = "path/to/tiddlywiki")
#> Error in make_run_txt(tidepath = "customPath", mws = TRUE, tiddlywiki = "path/to/tiddlywiki"): could not find function "make_run_txt"