Import and Run Tides

This function imports and runs the tides in your TiddlyWiki environment. It retrieves the tides and their associated scripts, and executes them in order, with an option to force run scripts tagged with DoNotRun.

Usage

run_tides(force_run = FALSE, mws = FALSE)

Arguments

force_run

Logical. If TRUE, also runs scripts tagged with DoNotRun. Default is FALSE.

mws

Logical. If TRUE, sets the TiddlyWiki options to use the "tides" recipe for a MultiWikiServer. Default is FALSE.

Examples

# Run tides with default settings
run_tides()
#> Error in curl::curl_fetch_memory(url, handle = handle): Failed to connect to 127.0.0.1 port 8080 after 0 ms: Connection refused

# Run tides and force scripts tagged with `DoNotRun` to run
run_tides(force_run = TRUE)
#> Error in curl::curl_fetch_memory(url, handle = handle): Failed to connect to 127.0.0.1 port 8080 after 0 ms: Connection refused

# Run tides in MultiWikiServer mode
run_tides(mws = TRUE)
#> Error in curl::curl_fetch_memory(url, handle = handle): Failed to connect to 127.0.0.1 port 8080 after 0 ms: Connection refused